Fold langsmith adapter into importer package
