+++ plugins/README.md
@@ -13,8 +13,7 @@ Kitaru keeps the default catalog in `src/kitaru/server/api/bootstrap.py`. At sta
 | `jsonl-importer` | `kitaru-jsonl-importer` | Kitaru JSONL importer |
 | `langfuse-importer` | `kitaru-langfuse-importer` | Langfuse importer and importer-backed adapter |
 | `langgraph` | `kitaru-langgraph` | LangGraph recording and replay adapter |
-| `langsmith` | `kitaru-langsmith` | LangSmith importer-backed adapter |
-| `langsmith-importer` | `kitaru-langsmith-importer` | LangSmith importer |
+| `langsmith-importer` | `kitaru-langsmith-importer` | LangSmith importer and importer-backed adapter |
 | `logfire` | `kitaru-logfire` | Logfire importer-backed adapter |
 | `logfire-importer` | `kitaru-logfire-importer` | Logfire importer |
 | `openai-agents` | `kitaru-openai-agents` | OpenAI Agents SDK recording adapter |
