+++ plugins/packages/langsmith-importer/pyproject.toml
@@ -22,7 +22,10 @@ classifiers = [
     "Topic :: Software Development :: Libraries :: Python Modules",
 ]
 keywords = ["ai-agents", "kitaru", "langsmith", "observability", "traces"]
-dependencies = ["kitaru>=0.22.0"]
+dependencies = [
+    "kitaru>=0.23.0",
+    "langsmith>=0.3,<1",
+]
 
 [project.urls]
 Homepage = "https://kitaru.ai"
