+++ plugins/packages/langsmith-importer/src/kitaru_langsmith_importer/__init__.py
@@ -1,1 +1,7 @@
 """LangSmith trace importer for Kitaru."""
+
+from .adapter import LangSmithAdapter
+
+__all__ = [
+    "LangSmithAdapter",
+]
