+++ plugins/packages/langsmith-importer/src/kitaru_langsmith_importer/adapter.py
@@ -14,7 +14,6 @@
 """LangSmith importer-backed adapter."""
 
 import asyncio
-import json
 import uuid
 from collections.abc import Iterator
 from contextlib import contextmanager
@@ -25,31 +24,13 @@ from langsmith.run_trees import get_cached_client
 from langsmith.schemas import Run
 
 from kitaru.importer_adapter import ImporterBackedAdapter
-from kitaru_langsmith_importer.importer import parse
+
+from .api import fetch_runs, serialize_runs, wait_for_runs
+from .importer import parse
 
 __all__ = ["LangSmithAdapter"]
 
-_POLL_INTERVAL = 2.0
 _ROOT_RUN_NAME = "kitaru-run"
-
-
-def _list_runs(client: Client, trace_id: str) -> list[Run]:
-    """Fetch all runs of one trace via the LangSmith API.
-
-    Args:
-        client: LangSmith client.
-        trace_id: LangSmith trace id.
-
-    Returns:
-        Trace runs.
-    """
-    return list(client.list_runs(trace_id=trace_id))
-
-
-def _trace_has_ended(runs: list[Run], trace_id: str) -> bool:
-    """Return whether the root run is present and every run has an end time."""
-    has_root = any(str(run.id) == trace_id for run in runs)
-    return has_root and all(run.end_time is not None for run in runs)
 
 
 class LangSmithAdapter(ImporterBackedAdapter):
@@ -97,17 +78,7 @@ class LangSmithAdapter(ImporterBackedAdapter):
         # Flush in a worker thread because the SDK call blocks on network
         # delivery.
         await asyncio.to_thread(client.flush)
-        # The trace is complete when the root run is present, every run has
-        # an end time, and the run count is stable across two consecutive
-        # polls.
-        previous_count: int | None = None
-        while True:
-            runs = await asyncio.to_thread(_list_runs, client, external_id)
-            if len(runs) == previous_count and _trace_has_ended(runs, external_id):
-                self._completed_runs[external_id] = runs
-                return
-            previous_count = len(runs)
-            await asyncio.sleep(_POLL_INTERVAL)
+        self._completed_runs[external_id] = await wait_for_runs(client, external_id)
 
     async def fetch(self, external_id: str) -> bytes:
         """Fetch the finished trace as LangSmith run JSONL.
@@ -121,9 +92,5 @@ class LangSmithAdapter(ImporterBackedAdapter):
         runs = self._completed_runs.pop(external_id, None)
         if runs is None:
             client = self._client or get_cached_client()
-            runs = await asyncio.to_thread(_list_runs, client, external_id)
-        # Serialize with the snake_case run field names the importer parser
-        # expects.
-        return "\n".join(
-            json.dumps(run.model_dump(mode="json")) for run in runs
-        ).encode("utf-8")
+            runs = await fetch_runs(client, external_id)
+        return serialize_runs(runs)
