--- plugins/packages/langsmith/CHANGELOG.md
@@ -1,5 +1,0 @@
-# Changelog
-
-## 0.1.0
-
-- Initial release of the LangSmith importer-backed adapter.
