+++ tests/scripts/test_release_units.py
@@ -32,7 +32,6 @@ EXPECTED_UNITS = {
     "langgraph": "kitaru-langgraph",
     "logfire": "kitaru-logfire",
     "logfire-importer": "kitaru-logfire-importer",
-    "langsmith": "kitaru-langsmith",
     "langsmith-importer": "kitaru-langsmith-importer",
     "openai-agents": "kitaru-openai-agents",
     "phoenix": "kitaru-phoenix",
@@ -119,7 +118,7 @@ def core_release_repo(
     return tmp_path, version
 
 
-def test_inventory_describes_core_and_thirteen_plugin_distributions() -> None:
+def test_inventory_describes_core_and_twelve_plugin_distributions() -> None:
     inventory = load_inventory()
 
     assert {unit.slug: unit.distribution for unit in inventory.units} == EXPECTED_UNITS
@@ -643,7 +642,7 @@ def test_text_and_json_outputs_contain_the_same_unit_identities() -> None:
     assert all(unit.distribution in text_output for unit in inventory.units)
 
 
-def test_plugin_matrix_is_generated_from_the_thirteen_plugin_units() -> None:
+def test_plugin_matrix_is_generated_from_the_twelve_plugin_units() -> None:
     matrix = build_plugin_matrix(load_inventory())
 
     assert matrix == {
@@ -833,7 +832,7 @@ def _run_cli(*arguments: str) -> subprocess.CompletedProcess[str]:
     [
         (["list"], "SLUG\tDISTRIBUTION\tVERSION\tDEFAULT\tTAG"),
         (["resolve", "--unit", "kitaru"], "python/kitaru/v"),
-        (["validate"], "Validated 14 release units."),
+        (["validate"], "Validated 13 release units."),
         (
             [
                 "propose-core-version",
